Move main run function into module folder
