+++ src/xia2/Modules/SSX/xia2_ssx.py
@@ -1,0 +1,232 @@
+from __future__ import annotations
+
+import logging
+import os
+import pathlib
+
+import iotbx.phil
+from libtbx import Auto
+from libtbx.introspection import number_of_processors
+
+from xia2.Modules.SSX.data_integration_programs import (
+    IndexingParams,
+    IntegrationParams,
+    RefinementParams,
+    SpotfindingParams,
+)
+from xia2.Modules.SSX.data_integration_standard import (
+    AlgorithmParams,
+    FileInput,
+    run_data_integration,
+)
+from xia2.Modules.SSX.util import report_timing
+
+# from xia2.Modules.SSX.data_reduction_simple import (
+#    SimpleDataReduction,
+#    SimpleReductionParams,
+# )
+
+
+xia2_logger = logging.getLogger(__name__)
+
+phil_str = """
+image = None
+  .type = str
+  .multiple = True
+  .help = "Path to image files"
+template = None
+  .type = str
+  .help = "The image sequence template"
+  .multiple = True
+directory = None
+  .type = str
+  .help = "A directory with images"
+  .multiple = True
+mask = None
+  .type = str
+  .help = "A mask to use for spotfinding and integration"
+reference_geometry = None
+  .type = path
+  .help = "Path to reference geomtry .expt file"
+nproc = Auto
+  .type = int
+space_group = None
+  .type = space_group
+  .help = "Space group to be used for indexing and integration."
+  .expert_level = 0
+d_min = None
+  .type = float
+  .help = "Resolution cutoff for spotfinding and integration."
+batch_size = 1000
+  .type = int
+  .help = "Index and integrate the images in batches with at least this number"
+          "of images, with a subfolder for each batch. This is a means to manage"
+          "the resource requirements and output reporting of the program, but"
+          "does not change the resultant integrated data."
+spotfinding {
+  min_spot_size = 3
+    .type = int
+    .help = "The minimum spot size to allow in spotfinding."
+  max_spot_size = 20
+    .type = int
+    .help = "The maximum spot size to allow in spotfinding."
+}
+indexing {
+  unit_cell = None
+    .type = unit_cell
+  max_lattices = 1
+    .type = int
+    .help = "Maximum number of lattices to search for, per image"
+}
+integration {
+  algorithm = stills *ellipsoid
+    .type = choice
+  ellipsoid {
+    rlp_mosaicity = *angular4 angular2 simple1 simple6
+      .type = choice
+  }
+}
+"""
+
+workflow_phil = """
+assess_crystals {
+  n_images = 1000
+    .type = int(value_min=1)
+    .help = "Number of images to use for crystal assessment."
+  images_to_use = None
+    .type = str
+    .help = "Specify an inclusive image range to use for crystal assessment,"
+            "in the form start:end"
+}
+geometry_refinement {
+  n_images = 1000
+    .type = int(value_min=1)
+    .help = "Number of images to use for reference geometry determination."
+  images_to_use = None
+    .type = str
+    .help = "Specify an inclusive image range to use for reference geometry"
+            "determination, in the form start:end"
+  outlier.algorithm = null auto mcd tukey *sauter_poon
+    .help = "Outlier rejection algorithm for joint refinement. If auto is"
+            "selected, the algorithm is chosen automatically."
+    .type = choice
+}
+workflow {
+  stop_after_geometry_refinement = False
+    .help = "If True, only perform spotfinding, indexing and joint refinement."
+    .type = bool
+  stop_after_integration = True
+    .help = "If True, do not perform data reduction after data integration."
+    .type = bool
+}
+"""
+
+# full_phil_str = phil_str + data_reduction_phil_str + workflow_phil
+full_phil_str = phil_str + workflow_phil
+
+
+@report_timing
+def run_xia2_ssx(
+    root_working_directory: pathlib.Path, params: iotbx.phil.scope_extract
+) -> None:
+    """
+    Run data integration for ssx images.
+    """
+    # First separate out some of the input params into relevant sections.
+    # Although it would be nice to have these as lists of Paths, we will need
+    # to compare to json data, add them as command line arguments etc, so simpler
+    # to have as strings from the start.
+    file_input = FileInput()
+    if params.image:
+        file_input.images = [str(pathlib.Path(i).resolve()) for i in params.image]
+    elif params.template:
+        file_input.templates = [str(pathlib.Path(i).resolve()) for i in params.template]
+    elif params.directory:
+        file_input.directories = [
+            str(pathlib.Path(i).resolve()) for i in params.directory
+        ]
+    else:
+        raise ValueError(
+            "No input data identified (use images=, template= or directory=)"
+        )
+    if params.mask:
+        file_input.mask = pathlib.Path(params.mask).resolve()
+    if params.reference_geometry:
+        reference = pathlib.Path(params.reference_geometry).resolve()
+        if not reference.is_file():
+            xia2_logger.warn(
+                f"Unable to find reference geometry at {os.fspath(reference)}, proceeding without this reference"
+            )
+        else:
+            file_input.reference_geometry = reference
+
+    options = AlgorithmParams(
+        batch_size=params.batch_size,
+        stop_after_geometry_refinement=params.workflow.stop_after_geometry_refinement,
+    )
+
+    if params.assess_crystals.images_to_use:
+        if ":" not in params.assess_crystals.images_to_use:
+            raise ValueError("Images to use must be given in format start:end")
+        vals = params.assess_crystals.images_to_use.split(":")
+        if len(vals) != 2:
+            raise ValueError("Images to use must be given in format start:end")
+        start = min(0, int(vals[0]) - 1)  # convert from image number to slice
+        end = int(vals[1])
+        options.assess_images_to_use = (start, end)
+    else:
+        options.assess_images_to_use = (0, params.assess_crystals.n_images)
+
+    if params.geometry_refinement.images_to_use:
+        if ":" not in params.geometry_refinement.images_to_use:
+            raise ValueError("Images to use must be given in format start:end")
+        vals = params.geometry_refinement.images_to_use.split(":")
+        if len(vals) != 2:
+            raise ValueError("Images to use must be given in format start:end")
+        start = min(0, int(vals[0]) - 1)  # convert from image number to slice
+        end = int(vals[1])
+        options.refinement_images_to_use = (start, end)
+    else:
+        options.refinement_images_to_use = (0, params.geometry_refinement.n_images)
+
+    if params.nproc is Auto:
+        params.nproc = number_of_processors(return_value_if_unknown=1)
+
+    spotfinding_params = SpotfindingParams(
+        params.spotfinding.min_spot_size,
+        params.spotfinding.max_spot_size,
+        params.d_min,
+        params.nproc,
+    )
+    indexing_params = IndexingParams(
+        params.space_group,
+        params.indexing.unit_cell,
+        params.indexing.max_lattices,
+        params.nproc,
+    )
+    refinement_params = RefinementParams(params.geometry_refinement.outlier.algorithm)
+    integration_params = IntegrationParams(
+        params.integration.algorithm,
+        params.integration.ellipsoid.rlp_mosaicity,
+        params.d_min,
+        params.nproc,
+    )
+
+    processed_batch_directories = run_data_integration(
+        root_working_directory,
+        file_input,
+        options,
+        spotfinding_params,
+        indexing_params,
+        refinement_params,
+        integration_params,
+    )
+    if not processed_batch_directories or params.workflow.stop_after_integration:
+        return
+
+    # Now do the data reduction
+    # if not params.symmetry.space_group:
+    #    params.symmetry.space_group = params.space_group
+    # reduction_params = SimpleReductionParams.from_phil(params)
+    # reducer = SimpleDataReduction(root_working_directory, processed_batch_directories)
+    # reducer.run(reduction_params)
